pyproject.toml
[project]
name = "portmgr"
version = "1.8.0dev2"
description = "Simple command interface to manage multiple Docker container"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "humanfriendly~=10.0",
    "jsonschema~=3.2.0",
    "pyyaml~=6.0.2",
    "tabulate~=0.8.9",
]
[tool.uv]
package = true
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[project.urls]
Website = "https://github.com/Craeckie/portmgr"
Repository = "https://github.com/Craeckie/portmgr.git"
[project.scripts]
portmgr = "portmgr:main"
